requirements_dev.txt
Pillow >= 6.2.1
pytest-django
pytest-cov
psycopg2-binary
psycopg[binary]
flake8
